+++ config_resolver/util.py
@@ -4,9 +4,11 @@ Helpers and utilities for the config_resolver package.
 This module contains stuff which is not directly impacting the business logic of
 the config_resolver package.
 """
+from logging import Filter, LogRecord
+from typing import Any
 
 
-class PrefixFilter(object):
+class PrefixFilter(Filter):
     """
     A logging filter which prefixes each message with a given text.
 
@@ -16,11 +18,12 @@ class PrefixFilter(object):
     """
     # pylint: disable = too-few-public-methods
 
-    def __init__(self, prefix, separator=' '):
+    def __init__(self, prefix: str, separator: str = ' ') -> None:
+        super().__init__()
         self._prefix = prefix
         self._separator = separator
 
-    def __eq__(self, other):
+    def __eq__(self, other: Any) -> bool:
         # NOTE: using ``isinstance(other, PrefixFilter)`` did NOT work properly
         # when running the unit-tests through ``sniffer``. Does this have
         # something to do with ``sniffer`` or is there something wrong with the
@@ -30,15 +33,15 @@ class PrefixFilter(object):
         # ``_separator`` member. They would wrongly be assumed to be the same.
         # I'll assume this won't happen for now.
         # pylint: disable = protected-access
-        return (self.__class__.__name__ == other.__class__.__name__ and
+        return (self.__class__.__name__ == other.__class__.__name__ and  # type: ignore
                 other._prefix == self._prefix and
                 other._separator == self._separator)
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return 'PrefixFilter(prefix={!r}, separator={!r}>'.format(
             self._prefix, self._separator)
 
-    def filter(self, record):
+    def filter(self, record: LogRecord) -> bool:
         # pylint: disable = missing-docstring
         record.msg = self._separator.join([self._prefix, record.msg])
         return True